moved constraint foldout drawing into utility method
