+++ Assets/MRTK/SDK/Experimental/Editor/Inspectors/BoundsControl/BoundsControlInspector.cs
@@ -9,7 +9,6 @@ using Microsoft.MixedReality.Toolkit.Utilities.Editor;
 using UnityEditor;
 using UnityEngine;
 using Microsoft.MixedReality.Toolkit.UI;
-using System.Linq;
 
 namespace Microsoft.MixedReality.Toolkit.Experimental.Inspectors
 {
@@ -135,43 +134,7 @@ namespace Microsoft.MixedReality.Toolkit.Experimental.Inspectors
                     }
 
                     EditorGUILayout.Space();
-                    constraintsFoldout = EditorGUILayout.Foldout(constraintsFoldout, "Constraints", true);
-
-                    if (constraintsFoldout)
-                    {
-                        if (EditorGUILayout.DropdownButton(new GUIContent("Add Constraint"), FocusType.Keyboard))
-                        {
-                            // create the menu and add items to it
-                            GenericMenu menu = new GenericMenu();
-
-                            var type = typeof(TransformConstraint);
-                            var types = System.AppDomain.CurrentDomain.GetAssemblies()
-                                        .SelectMany(s => s.GetLoadableTypes())
-                                        .Where(p => type.IsAssignableFrom(p) && !p.IsAbstract);
-
-                            foreach (var derivedType in types)
-                            {
-                                menu.AddItem(new GUIContent(derivedType.Name), false, t => boundsControl.gameObject.AddComponent((System.Type)t), derivedType);
-                            }
-
-                            menu.ShowAsContext();
-                        }
-
-                        var constraints = boundsControl.GetComponents<TransformConstraint>();
-
-                        foreach (var constraint in constraints)
-                        {
-                            EditorGUILayout.BeginHorizontal();
-                            string constraintName = constraint.GetType().Name;
-                            EditorGUILayout.LabelField(constraintName);
-                            if (GUILayout.Button("Go to component"))
-                            {
-                                Highlighter.Highlight("Inspector", $"{ObjectNames.NicifyVariableName(constraintName)} (Script)");
-                                EditorGUIUtility.ExitGUI();
-                            }
-                            EditorGUILayout.EndHorizontal();
-                        }
-                    }
+                    constraintsFoldout = InspectorUIUtility.DrawComponentTypeFoldout<TransformConstraint>(boundsControl.gameObject, constraintsFoldout, "Constraint");
 
                     EditorGUILayout.Space();
                     EditorGUILayout.LabelField(new GUIContent("Events", "Bounds Control Events"), EditorStyles.boldLabel, GUILayout.ExpandWidth(true));
